http://tintin.cs.indiana.edu:8091/browse/PFANDRES-28
made some changes to getCallInfo() to extract data. Instead of clicking
on the button, it takes the information fed to the javascript which is
used to build the table in the pop up.
added a pay field to event information to hold the total pay for the dat
that includes extra pay, tips, and rate*hours worked.